Use jwt extended

+++ resources/item.py
@@ -1,6 +1,12 @@
 import sqlite3
 from flask_restful import Resource, reqparse
-from flask_jwt import jwt_required, current_identity
+from flask_jwt_extended import (
+	jwt_required,
+	get_jwt_claims,
+	jwt_optional,
+	get_jwt_identity,
+	fresh_jwt_required
+)
 from models.item import ItemModel
 
 class Item(Resource):
@@ -16,16 +22,18 @@ class Item(Resource):
 		help="Every item needs a store id"
 	)
 
-	@jwt_required()
+	@jwt_required
 	def get(self, name):
 		item = ItemModel.find_by_name(name)
 		if item:
 			return item.json()
 		return {'message': 'Item not found'}, 404
 
+	# needs fresh access token
+	@fresh_jwt_required
 	def post(self,name):
 		if ItemModel.find_by_name(name):
-			return {'message': "An item with name '{}' already exsits.".format(name)}, 400
+			return {'message': "An item with name '{}' already exists.".format(name)}, 400
 
 		data = Item.parser.parse_args()
 		item = ItemModel(name, **data)
@@ -36,7 +44,12 @@ class Item(Resource):
 
 		return item.json() , 201
 
+	@jwt_required
 	def delete(self, name):
+		claims = get_jwt_claims()
+		if not claims['is_admin']:
+			return {'message': 'You need admin privileges for this action'}
+
 		item =  ItemModel.find_by_name(name)
 		if item:
 			item.delete_from_db()
@@ -58,5 +71,14 @@ class Item(Resource):
 
 
 class ItemList(Resource):
+	@jwt_optional
 	def get(self):
-		return {'items': list(map(lambda x : x.json(), ItemModel.query.all()))}
+		user_id = get_jwt_identity()
+		items = list(map(lambda x : x.json(), ItemModel.find_all()))
+		if user_id:
+			return {'items': items}, 200
+		return{
+			'items': [item['name'] for item in items],
+			'message': 'More data available if you log in'
+		}
+		#return {'items': list(map(lambda x : x.json(), ItemModel.find_all()))}
